pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crocus-tools"
version = "0.1.0"
description = "CROCUS Python tools for urban CFD preprocessing"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "geopandas",
    "laspy",
    "PyYAML",
    "shapely",
    "pyproj",
    "scipy",
    "pyvista",
    "matplotlib",
    "utm",
]

[project.scripts]
crocus = "crocus.cli:main"

[tool.setuptools.packages.find]
include = ["crocus*"]
exclude = ["crocus.tests*"]
